planning-agent/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-socketio>=5.11
openai>=1.40
python-dotenv>=1.0
pydantic>=2.6
httpx>=0.27
